+++ service/pipelines/model.go
@@ -418,6 +418,8 @@ type GetUpdateResponse struct {
 }
 
 type IngestionConfig struct {
+	// Select tables from a specific source report.
+	Report *ReportSpec `json:"report,omitempty"`
 	// Select tables from a specific source schema.
 	Schema *SchemaSpec `json:"schema,omitempty"`
 	// Select tables from a specific source table.
@@ -1273,6 +1275,32 @@ type PipelineTrigger struct {
 	Manual *ManualTrigger `json:"manual,omitempty"`
 }
 
+type ReportSpec struct {
+	// Required. Destination catalog to store table.
+	DestinationCatalog string `json:"destination_catalog,omitempty"`
+	// Required. Destination schema to store table.
+	DestinationSchema string `json:"destination_schema,omitempty"`
+	// Required. Destination table name. The pipeline fails if a table with that
+	// name already exists.
+	DestinationTable string `json:"destination_table,omitempty"`
+	// Required. Report URL in the source system.
+	SourceUrl string `json:"source_url,omitempty"`
+	// Configuration settings to control the ingestion of tables. These settings
+	// override the table_configuration defined in the
+	// IngestionPipelineDefinition object.
+	TableConfiguration *TableSpecificConfig `json:"table_configuration,omitempty"`
+
+	ForceSendFields []string `json:"-"`
+}
+
+func (s *ReportSpec) UnmarshalJSON(b []byte) error {
+	return marshal.Unmarshal(b, s)
+}
+
+func (s ReportSpec) MarshalJSON() ([]byte, error) {
+	return marshal.Marshal(s)
+}
+
 type SchemaSpec struct {
 	// Required. Destination catalog to store tables.
 	DestinationCatalog string `json:"destination_catalog,omitempty"`
@@ -1451,7 +1479,7 @@ type TableSpec struct {
 	DestinationCatalog string `json:"destination_catalog,omitempty"`
 	// Required. Destination schema to store table.
 	DestinationSchema string `json:"destination_schema,omitempty"`
-	// Optional. Destination table name. The pipeline fails If a table with that
+	// Optional. Destination table name. The pipeline fails if a table with that
 	// name already exists. If not set, the source table name is used.
 	DestinationTable string `json:"destination_table,omitempty"`
 	// Source catalog name. Might be optional depending on the type of source.
@@ -1485,6 +1513,10 @@ type TableSpecificConfig struct {
 	SalesforceIncludeFormulaFields bool `json:"salesforce_include_formula_fields,omitempty"`
 	// The SCD type to use to ingest the table.
 	ScdType TableSpecificConfigScdType `json:"scd_type,omitempty"`
+	// The column names specifying the logical order of events in the source
+	// data. Delta Live Tables uses this sequencing to handle change events that
+	// arrive out of order.
+	SequenceBy []string `json:"sequence_by,omitempty"`
 
 	ForceSendFields []string `json:"-"`
 }